Introduce Docker Compose env file
The current setup instructions require environment variables to be change in the
the Docker Compose file, but this is very error-prone. An env file is
recommended for managing environment variables.

An example file ('env.example') is provided, with default values, which should
be copied as '.env' in the same directory as the Docker Compose file, and
modified locally.

The setup instructions have been updated accordingly. First-time users are asked
to focus on the contents of the .env file and not interfere with the main Docker
Compose file. Additional instructions can be provided in the env file to give
more help about each env var.